requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Deprecated==1.2.18
Flask==3.1.2
flask-cors==6.0.1
Flask-Limiter==3.12
future==1.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.2
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
rich==13.9.4
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
yt-dlp==2025.8.20
